renamed pages folder to screens
