+++ App Programmering/pets/App.js
@@ -3,8 +3,8 @@ import { StyleSheet, Text, View } from 'react-native';
 import { createAppContainer } from 'react-navigation';
 
 import { createStackNavigator } from 'react-navigation-stack';
-import HomeScreen from './pages/HomeScreen';
-import PetView from './pages/PetView';
+import HomeScreen from './screens/HomeScreen';
+import PetView from './screens/PetView';
 
 //import { AppNavigator } from './components/Navigation';
